+++ src/eca/shared.clj
@@ -38,3 +38,5 @@
          (throw (IllegalArgumentException.
                  "assoc-some expects even number of arguments after map/vector, found odd number")))
        ret))))
+
+(defn multi-str [& strings] (string/join "\n" strings))