+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.9.0
+
 - Include eca as a  server with tools.
 - Support disable tools via config.
 - Improve ECA prompt to be more precise and output with better quality